Shift wallet and resources to inventory
What?
=====

This is the start of shifting resource management and a wallet (which
holds available currency) behind an `Inventory` type.

There are three discrete types of actions to be taken on an inventory:

* accrual of value based on a particular time period
* proactive generation of currency via user input (clicking)
* purchasing of resources

Additional game state/complexity (e.g. purchased static bonuses or
multipliers) would also be captured here in the future.