Calculate price based on the number of items purchased
What?
=====

This converts the inventory's resources from a list of resources to an
`AllDict.AllDict` with a `Resource.Level` as the key. The notion of having a
level as a key guarantees some amount of uniqueness across the various
resources and allows for simplification of the inventory due to the `Level`
being the identifier instead of the `Resource` (which changes over time due to
it currently tracking the total number purchased).