Generate random events
This introduces the concept of events. Every 15 seconds, there's
a 50% chance that a random event triggers. Events last for 10 seconds
before disappearing.

For now, events have no effect on the game. They are simply cosmetic. In
a future commit, we want the user to be able to interact with events in
ways that could affect the game.