Add cost discounts for resources
Why:

We want to be able to apply discounts to the cost of resources. These
could created from various places, like random events.

How:

This commit adds a discounts field to the inventory model which is a
list of limited multipliers. During the purchase of a resource a new
function Inventory.applyDiscounts will combine all current discounts and
update the final cost of the transaction. DecreaseGlobalCost and
DecreaseLevelCost where added the Multiplier type to express
multipliers that discount a resource cost.