Extract Data.GameConfiguration
What?
=====

This extracts names, values, and icons to a separate module in order to isolate
in one area all of the game configuration. As a result, modifying multipliers,
adding additional resource levels, and modifying the underlying mathematical
mechanics of the game will occur in one spot.

In theory, this could be extracted further and initialized when the application
starts, and based on a configuration setting, modified in real-time to test out
various game states.