Automatically persist game state
Why:

We would like the game to auto-save and load from localStorage on app
initialization.

This PR:

Introduces the wiring up communication between elm and JavaScript via
ports using the elm-localstorage package. It also introduces some JSON
encoders and decoders to be able to save the game inventory data to
localStorage. Finally we added a subscription to save every 5 seconds
and a side effect to load from localStorage on app initialization.

Co-authored-by: MattMSumner <matt.m.sumner@gmail.com>
Co-authored-by: JoelQ <joel.quen@gmail.com>
Co-authored-by: joshuaclayton <joshua.clayton@gmail.com>