Continue to accrue value when the browser tab isn't active
What?
=====

Certain browsers may reduce frequency of subscriptions being run when
the tab isn't active; this results in incorrect accrual of value over
time when the browser is in the background.

This introduces tracking of the last accrued time and compares it
against game update frequency; once the updated time passes outside a
10% threshold, the game uses that time distance to calculate accrued
value instead of the expected update frequency.