Expirable handles ticks even if ticks don't come at the correct times
What?
=====

Due to the nature of Elm subscriptions and background tabs, `Time.every`
doesn't actually guarantee commands get run at the appropriate interval.

In order to account for this, Expirable now keeps track of the last time
the event was run, and ensures it correctly decrements the seconds when
time has passed.

It should be noted that this can't actually guarantee the application
works correctly when in a background tab; values might be extracted from
expirables that technically shouldn't be there. What this does handle,
however, is the removal of the expirable next time the subscription
fires.